Envio de mensaje y recepción de ACK correcta por Modulo LoRa.

+++ Modules/RFTransciever/RFTranscieverState/RFTransicieverState.h
@@ -22,7 +22,7 @@ struct RFPacketMessage;
 class RFTransicieverState {
 public:
 //=====[Declaration of public methods]=========================================
-    virtual void addRFFormatToMessage (int deviceId, int messageNumber, char * messageToBeSend);
+    virtual void addRFFormatToMessage (long long int deviceId, int messageNumber, char * messageToBeSend);
     virtual void sendMessage (LoRaClass * LoRaModule, char * messageToBeSend, NonBlockingDelay * backoffTime);
     virtual bool getAcknowledgement (LoRaClass * LoRaModule, char * messageRecieved, NonBlockingDelay * timeOut);
 //=====[Declaration of privates atributes]=========================================
